requirements.txt
streamlit
pillow
requests
google-generativeai